Extract user handlers to separate module
