+++ src/handlers/index.js
@@ -2,4 +2,5 @@
 module.exports = {
     ...require('./handlers'),
     ...require('./tournaments.handlers'),
+    ...require('./users.handlers'),
 };
